Master-Detail layout part 3 - Simplify Detail activity - get it out of the business of updating the tab view.  Simplify TabContainerFragment too.  The fragments in the tab view now use Observer to watch for their own updates.  DetailActivity does not need to listen for data changes. Any change of movie is delivered by intent to the activity.
MovieDataService now has two entrances to the Singleton.  The one used by Activities will set the movie and kick of the data gathering.  The simpler read-only entrance is used by Fragments.

In other news, DetailFragment now pays more attention to the simple Movie while it waits for the more involved MovieDetail. This lets parts of the UI fill in faster for better perceived performance.  The initialization is now performed once, but in several steps - simple Movie, Movie Detail, then the background.  Previously it was running through the UI multiple times for the same data, which did nothing but slow things down.  Make sure to clear out the UI when the movies change, which is needed for Master-Detail.

Almost there for Master-Detail.  The UI is behaving as expected in single pane mode through multiple changes of movie, rotations, tab changes, etc. There is a glitch with the Empty Review message showing up when there are late-arriving reviews.