The reason why this was running out of memory and throwing odd errors is because Picasso was getting called from different fragments with different contexts.  Solution: use Picasso as a singleton with the Application context.  The garbage collector is not running constantly now.  I may want to add the largeHeap=true to the manifest though.
Also, we don't need to keep an extra, stale copy of the movies around in MainFragment.mMovieList.
MovieAdapter is keeping the real, up to date list of movies.  Get rid of mMovieList.
Also, the movies are surviving in MovieAdapter between rotations / fragment changes.  Don't reset the whole adapter if the sort type hasn't changed.

Still, this seems slower.  The Review callback is getting called a lot. I'll look at that next.